Minor refactor and table entry filters.
Fixed table entry filtering scheme that was causing some of the loop
examples to fail. Table entries are now first grouped by line number
which is also how they are collapsed, this seems to have broken a
different example, and the issue with closures still remains. Needed
still is documentation and addressing comments.
